Relocate firebase.ts and mongodb.ts to config folder under lib folder
