+++ app/components/ChangePassword.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { useState } from "react";
-import { auth } from "@/lib/firebase";
+import { auth } from "@/lib/configs/firebase";
 import {
   EmailAuthProvider,
   reauthenticateWithCredential,
